table-struct: Adding better naming on inline struct
