+++ example/command/console-script-table-json.go
@@ -13,7 +13,11 @@ type Author struct {
 type Book struct {
 	ISBN   *string
 	Title  *string
-	Author *Author `header:"inline"`
+	Author *Author `header:"Author"`
+}
+
+func (b Book) String() string {
+	return *b.ISBN
 }
 
 func ptrStr(str string) *string {
@@ -28,6 +32,17 @@ func main() {
 		{ISBN: ptrStr("960-425-059-0"), Title: ptrStr("The Lord of the Rings"), Author: &Author{Name: "J. R. R. Tolkien", Age: 81}},
 		{ISBN: ptrStr("80-902734-1-6"), Title: ptrStr("And Then There Were None"), Author: &Author{Name: "Agatha Christie", Age: 85}},
 	}
+
+	//books := map[string][]Book{
+	//	"bookshelves 1": {
+	//		{ISBN: ptrStr("99921-58-10-7"), Title: ptrStr("The Divine Comedy"), Author: &Author{Name: "Dante Alighieri", Age: 56}},
+	//		{ISBN: ptrStr("9971-5-0210-0"), Title: ptrStr("A Tale of Two Cities"), Author: &Author{Name: "Charles Dickens", Age: 58}},
+	//	},
+	//	"bookshelves 2": {
+	//		{ISBN: ptrStr("960-425-059-0"), Title: ptrStr("The Lord of the Rings"), Author: &Author{Name: "J. R. R. Tolkien", Age: 81}},
+	//		{ISBN: ptrStr("80-902734-1-6"), Title: ptrStr("And Then There Were None"), Author: &Author{Name: "Agatha Christie", Age: 85}},
+	//	},
+	//}
 
 	tab := table.
 		NewTable().
